cd to /data before running tool.
The tool has to be run in /data since it expects `discord-registration.yaml` to be in the working directory. However, the bridge itself can't be run from /data because it expects to be able to load the config schema from path relative to the working directory. The tools don't seem to need this schema and seem to run fine from /data.