Fix typo with options.reg
The '-r' flag value is stored under the alias 'registration', but is later passed along to ToolsHelper.getToolDependencies() as 'options.reg' which is undefined. This causes the internal default of './discord-registration.yaml' to always be used, instead of what the user defined. Should be fixed by changing options.reg to options.registration.